refactor mcp generate base input prep
